refactor: separate msw handlers from graphql to testing and remove testing from bundle
