+++ packages/ui-kit/src/components/Leaderboard/Leaderboard.test.tsx
@@ -1,8 +1,8 @@
 import { render, waitFor } from '@testing-library/react'
 import { Chart } from 'chart.js'
 import React from 'react'
-import { mockLeaderboardQuery, RelativeTimeRange } from '../../helpers'
-import { Dom, setupTestHandlers } from '../../helpers/testing'
+import { RelativeTimeRange } from '../../helpers'
+import { Dom, setupTestHandlers, mockLeaderboardQuery } from '../../helpers/testing'
 import { Leaderboard } from './Leaderboard'
 
 const mockData = {
